# Upgrade to rand 0.6