--- a/Cargo.toml
+++ b/Cargo.toml
@@ -1,7 +1,7 @@
 [package]
 authors = ["Dylan MacKenzie <ecstaticmorse@gmail.com>"]
 name = "ksuid"
-version = "0.1.0"
+version = "0.3.0"
 description = "A library for efficiently generating, parsing and serializing Segment.io KSUIDs"
 repository = "https://github.com/ecstatic-morse/ksuid.git"
 readme = "README.md"
@@ -10,7 +10,7 @@ license = "MIT"
 
 [dependencies]
 byteorder = "1.0.0"
-rand = "0.3.15"
+rand = "0.6.0"
 resize-slice = "0.1.2"
 time = "0.1.37"