--- a/cli/Cargo.toml
+++ b/cli/Cargo.toml
@@ -1,7 +1,7 @@
 [package]
 authors = ["Dylan MacKenzie <ecstaticmorse.gmail.com>"]
 name = "ksuid-cli"
-version = "0.1.0"
+version = "0.3.0"
 description = "A command line interface for generating, parsing, and serializing Segment.io KSUIDs"
 repository = "https://github.com/ecstatic-morse/ksuid.git"
 readme = "../README.md"
@@ -10,8 +10,7 @@ license = "MIT"
 
 [dependencies]
 docopt = "0.8.0"
-ksuid = { version="0.1.0", path = ".." }
-rand = "0.3.15"
+ksuid = { version="0.3.0", path = ".." }
 serde = "1.0.8"
 serde_derive = "1.0.8"
 time = "0.1.37"