--- a/cli/src/main.rs
+++ b/cli/src/main.rs
@@ -4,13 +4,11 @@ extern crate serde_derive;
 extern crate docopt;
 extern crate ksuid;
 extern crate time;
-extern crate rand;
 
 use std::io::{self, Write};
 use std::process::exit;
 
 use ksuid::Ksuid;
-use rand::Rng;
 
 const USAGE: &str = "
 ksuid
@@ -46,10 +44,8 @@ fn generate(args: Args) {
     let out = io::stdout();
     let mut locked = out.lock();
 
-    let mut rng = rand::thread_rng();
-
     for _ in 0..args.flag_count {
-        writeln!(&mut locked, "{}", rng.gen::<Ksuid>().to_base62()).unwrap();
+        writeln!(&mut locked, "{}", Ksuid::generate().to_base62()).unwrap();
     }
 }