Update to latest release. Better performance as a side effect.

Before:

    running 2 tests
    test tests::bench_gen                         ... bench:         159 ns/iter (+/- 3)
    test tests::bench_gen_lock_rng                ... bench:         155 ns/iter (+/- 2)

After:

    running 2 tests
    test tests::bench_gen                         ... bench:          74 ns/iter (+/- 2)
    test tests::bench_gen_lock_rng                ... bench:          72 ns/iter (+/- 4)

Bump version to 0.3.0 because Ksuid no longer impl Rand

The ksuid cli now uses public `Ksuid::generate()` interface so it
doesn't need to import rand crate directly.